python_backend/requirements.txt
firebase-admin==6.2.0
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
apscheduler==3.10.1